feat: implemented raw record endpoing
