+++ api/app/db.py
@@ -7,9 +7,7 @@ TORTOISE_ORM = {
         "app": {
             "models": [
                 "aerich.models",
-                "app.models.raw",
-                "app.models.std",
-                "app.models.mrg"
+                "app.models",
             ],
             "default_connection": "default",
         },
