+++ api/app/utils.py
@@ -5,7 +5,7 @@ from jose import jwt
 from passlib.context import CryptContext
 
 from app import config
-from api.app.models.mrg import User
+from app.models import User
 
 pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
