+++ api/scripts/database_cleanup.py
@@ -3,7 +3,7 @@ from loguru import logger
 from tortoise import Tortoise, run_async
 from tortoise.models import Model
 
-from api.app.models import mrg
+from app import models
 from app.db import TORTOISE_ORM
 
 
@@ -12,7 +12,7 @@ async def run():
     await Tortoise.generate_schemas()
 
     # Clean up database
-    for _, v in mrg.__dict__.items():
+    for _, v in models.__dict__.items():
         if isinstance(v, type) and issubclass(v, Model) and v != Model:
             logger.info(f"Cleaning up {v.__name__}")
             await v.all().delete()
